refactor(transform): split custom containers into sibling modules
Avoid mod.rs and keep each new file under the 350-line CI limit.
